App/requirements.txt
flask
firebase_admin
pre-commit
pytest
python-dotenv
